+++ src/app/common/task.js
@@ -5,15 +5,18 @@ const yerbamate = require('yerbamate');
 
 const TaskStatus = require('./task_status');
 const {TaskTimer, InverseTaskTimer} = require('./task_timer');
+const utils = require('./utils');
+const constants = require('../../common/constants');
 
 const outputMaxSize = 6000;
 
 class Task {
-    constructor(title, path, command, env) {
-        this.title = title.trim() || "";
-        this.command = command || "";
-        this.path = path || "";
-        this.env = env || [];
+    constructor(options) {
+        this.title = options.title.trim() || "";
+        this.command = options.command || "";
+        this.description = this._formatDescription(options.description);
+        this.path = options.path || "";
+        this.env = options.env || [];
         this.status = TaskStatus.idle;
 
         this.output = null;
@@ -83,6 +86,7 @@ class Task {
         };
         if (this.env && this.env.length > 0) res.env = this.env;
         if (this.path !== "") res.path = this.path;
+        if (this.description !== "") res.description = this.description;
         return res;
     }
 
@@ -96,6 +100,15 @@ class Task {
                 this.run(done);
             }, seconds * 1000);
         });
+    }
+
+    clone() {
+        return new Task({
+            title: this.title,
+            description: this.description,
+            path: this.path,
+            command: this.command,
+            env: this.env});
     }
 
     _processCommand() {
@@ -121,6 +134,11 @@ class Task {
             return acc;
         }, {});
         return res;
+    }
+
+    _formatDescription(originalDescription) {
+        const description = originalDescription || "";
+        return utils.truncate(description.trim(), constants.maxDescriptionLength).trim();
     }
 }