+++ src/app/components/task/task_card.vue
@@ -1,32 +1,43 @@
 <template>
     <div class="task-card" @contextmenu.stop="context()">
         <div class="columns is-mobile task-card-header" @click="taskSelected">
-            <div class="column">
-                <p>
-                    <span v-if="editMode" class="drag-handle">
-                        <span class="icon">
-                            <i class="fas fa-equals"></i>
-                        </span>
-                    </span>
-                    <span v-show="open" class="icon">
-                        <i class="fas fa-caret-down"></i>
-                    </span>
-                    <span v-show="!open" class="icon">
-                        <i class="fas fa-caret-right"></i>
-                    </span>
-                    {{task.title}}
-                </p>
-            </div>
-            <div class="column">
+            <div class="column is-half">
                 <div class="columns is-mobile">
-                    <div v-if="showTimer" class="column has-text-centered">
-                        <p>{{executionTime}}</p>
+                    <div v-if="editMode" class="column is-narrow drag-handle-container">
+                        <p>
+                            <span class="drag-handle">
+                                <span class="icon">
+                                    <i class="fas fa-equals"></i>
+                                </span>
+                            </span>
+                        </p>
+                    </div>
+                    <div class="column">
+                        <b>
+                            <span v-show="open" class="icon">
+                                <i class="fas fa-caret-down"></i>
+                            </span>
+                            <span v-show="!open" class="icon">
+                                <i class="fas fa-caret-right"></i>
+                            </span>
+                            {{task.title}}
+                        </b>
+                        <p class="description" v-show="task.description">
+                            {{task.description}}
+                        </p>
+                    </div>
+                </div>
+            </div>
+            <div class="column is-half">
+                <div class="columns is-mobile task-actions">
+                    <div v-if="showTimer" class="column has-text-centered task-action-text">
+                        <p class="execution-time">{{executionTime}}</p>
                     </div>
                     <div class="column">
                         <button v-if="!editMode" :class="{'is-danger':running}" class="button is-primary task-button" @click.stop="toggleRun" :disabled="!runButtonEnabled">{{runButtonText}}</button>
                         <button v-else class="button is-primary task-button is-danger" @click.stop="deleteTask">Delete</button>
                     </div>
-                    <div class="column">
+                    <div class="column task-action-text">
                         <task-status :status="status"></task-status>
                     </div>
                 </div>
@@ -178,8 +189,15 @@ module.exports = {
     margin-top: 0;
 }
 
-p {
-    padding-top: 4px;
+.description {
+    font-size: 15px;
+    color: #9e9e9e;
+    margin-left: 42px;
+    padding-top: 0;
+}
+
+.execution-time{
+    padding-top: 4px
 }
 
 .task-card-header {
@@ -191,6 +209,20 @@ p {
     border-bottom-style: solid;
     border-bottom-width: 1px;
     border-color: #e2e2e2;
+    height: 72px;
+
+    .task-actions {
+        margin-top: -0.5rem;
+
+        .task-action-text {
+            padding-top: 1rem;
+        }
+    }
+}
+
+.drag-handle-container{
+    padding-right: 0;
+    padding-top: 1.5rem;
 }
 
 .task-output-container{