+++ src/common/constants.js
@@ -1,0 +1,7 @@
+"use strict";
+
+module.exports = {
+    maxTaskNameLength: 26,
+    maxDescriptionLength: 32,
+    maxSuiteNameLength: 23
+};